Refactor OutboundMessageHandler
